% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennvis.R
\name{calculateDistance}
\alias{calculateDistance}
\title{Distance calculation function}
\usage{
calculateDistance(rx, ry, cov, precision = 1000)
}
\description{
Distance calculation function
}
\keyword{internal}
