% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.vennvis}
\alias{plot.vennvis}
\title{Replot method}
\usage{
\method{plot}{vennvis}(x, ...)
}
\arguments{
\item{x}{a vennvis object returned from vennvis() call}

\item{...}{any other args}
}
\description{
Replot method
}
