% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.vennvis}
\alias{print.vennvis}
\title{Print method}
\usage{
\method{print}{vennvis}(x, ...)
}
\arguments{
\item{x}{a vennvis object returned from vennvis() call}

\item{...}{any other args}
}
\description{
Print method
}
