% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennvis.R
\name{vennvis}
\alias{vennvis}
\title{Visualise the relation between two variables as circles}
\usage{
vennvis(x, y, scale = FALSE, prec = 400, plot = TRUE, names = TRUE,
  points = FALSE)
}
\arguments{
\item{x}{a base variable (vector)}

\item{y}{a comparison variable (vector)}

\item{scale}{whether to scale the variables}

\item{prec}{the precision of the drawing and the numerical optimisation}

\item{plot}{whether to plot, default TRUE}

\item{names}{display the names of the variables, default TRUE}

\item{points}{plot informative points, default FALSE}
}
\description{
Using a venn diagram where the overlap indicates the covariance between two
variables and the area of the circles indicates the variance of each.
}
