% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennvis.R
\name{vennvis}
\alias{vennvis}
\title{Visualise the relation between two variables as circles}
\usage{
vennvis(x, y, z, scale = FALSE, plot = FALSE, precision = 400,
  axes = FALSE)
}
\arguments{
\item{x}{a base variable (vector)}

\item{y}{a comparison variable (vector)}

\item{scale}{whether to scale the variables}

\item{plot}{whether to plot, default TRUE}

\item{precision}{the precision of the drawing}
}
\description{
Using a venn diagram where the overlap indicates the covariance between two
variables and the area of the circles indicates the variance of each.
}
